fix undefined behavior shifting by 32 bits. make integer promotion explicit to silence warning. http://cr.bittorrent.com/6334973984374784/

+++ src/blockallocator.cpp
@@ -11,7 +11,7 @@ template<typename T> T static inline exch(T &a, const T b) {
 
 void BlockAllocator::Grow()
 {
-	byte *block = new byte[_size * _grow];
+	byte *block = new byte[int64(_size) * _grow];
 	for(int i=_grow; --i>=0; ) {
 		void *a = block + i * _size;
 		*(void**)a = exch(_free, a);
